asamblea_api/requirements.txt
fastapi==0.109.1
uvicorn==0.23.0
sqlalchemy==1.4.49
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
slowapi==0.1.9
prometheus-fastapi-instrumentator==7.0.0